% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recapture.uploading.R
\name{batch_upload_recaptures}
\alias{batch_upload_recaptures}
\title{batch_upload_recaptures}
\usage{
batch_upload_recaptures(
  db = NULL,
  backups = T,
  oracle.user = if (exists("oracle.lobtag.user", inherits = T)) oracle.lobtag.user else
    NULL,
  oracle.password = if (exists("oracle.lobtag.password", inherits = T))
    oracle.lobtag.password else NULL,
  oracle.dbname = if (exists("oracle.lobtag.server", inherits = T)) oracle.lobtag.server
    else NULL
)
}
\description{
batch uploads tag recaptures data
}
