% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.uploading.R
\name{clean_NAs}
\alias{clean_NAs}
\title{clean_NAs}
\usage{
clean_NAs(
  db = NULL,
  oracle.user = NULL,
  oracle.password = NULL,
  oracle.dbname = NULL,
  table.name = NULL,
  close.con = T
)
}
\description{
function for cleaning improperly stored NA (or NULL or blank) values in database (should be (null) in SQL tables, sometimes gets stored as character NA,NULL,"")
}
