% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{db_connection}
\alias{db_connection}
\title{db_connection}
\usage{
db_connection(
  db = db,
  oracle.user = oracle.user,
  oracle.password = oracle.password,
  oracle.dbname = oracle.dbname
)
}
\description{
establishes connection to databse of users choosing
}
