% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_deletion.R
\name{delete_recaptures}
\alias{delete_recaptures}
\title{delete_recaptures}
\usage{
delete_recaptures(
  db = NULL,
  oracle.user = if (exists("oracle.personal.user")) oracle.personal.user else NULL,
  oracle.password = if (exists("oracle.personal.password")) oracle.personal.password else
    NULL,
  oracle.dbname = if (exists("oracle.personal.server")) oracle.personal.server else NULL
)
}
\description{
allows user to delete chosen tag recaptures and associated paths
}
