% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.generation.R
\name{generate_maps}
\alias{generate_maps}
\title{generate_maps}
\usage{
generate_maps(
  map_token =
    "pk.eyJ1IjoiZWxlbWVudGpha2UiLCJhIjoiY2xxZmh2bGFiMHhyNTJqcHJ0cDBqcW83ZiJ9.mDd49ObNcNdG6MzH3yc2dw",
  person = NULL,
  output.location = "C:/Users/ELEMENTG/Documents/Rsaves"
)
}
\description{
creates maps of tag movement for participants
}
