% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.generation.R
\name{generate_paths}
\alias{generate_paths}
\title{generate_paths}
\usage{
generate_paths(
  username = oracle.personal.user,
  password = oracle.personal.password,
  dbname = oracle.personal.server,
  tags = "all",
  depth.raster.path = system.file("data", "depthraster2.tif", package = "LobTag2"),
  neighborhood = 16,
  type = "least.cost"
)
}
\description{
Uses releases and recapture data with spatial/depth information to draw plausible paths of animal movement
}
