% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobster.letters.R
\name{lobster.letters}
\alias{lobster.letters}
\title{lobster.letters}
\usage{
lobster.letters(
  people = NULL,
  db = "Oracle",
  only.unrewarded = T,
  output.location = NULL,
  oracle.user = oracle.personal.user,
  oracle.password = oracle.personal.password,
  oracle.dbname = oracle.personal.server
)
}
\description{
create reward letters for fishers who report tags, letters contain map and text.

Letter text for rewards markdown
}
