% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.generation.R
\name{map_by_factor}
\alias{map_by_factor}
\title{map_by_factor}
\usage{
map_by_factor(
  factor.from = NULL,
  map.by = NULL,
  filter.maps.for = NULL,
  group.by = NULL,
  all.releases = F,
  tag.prefix = NULL,
  add.paths = F,
  map.token = mapbox.token,
  db = NULL,
  output.location = NULL,
  oracle.user = oracle.personal.user,
  oracle.password = oracle.personal.password,
  oracle.dbname = oracle.personal.server,
  set.output = T,
  set.inset = T,
  max.pixels = 8e+05,
  map.res = 0.9,
  inset.map = T
)
}
\description{
specific mapping function for mapping releases and returns by custom factor
}
