% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recapture.uploading.R
\name{upload_recaptures}
\alias{upload_recaptures}
\title{upload_recaptures}
\usage{
upload_recaptures(
  db = NULL,
  backups = T,
  oracle.user = if (exists("oracle.personal.user")) oracle.personal.user else NULL,
  oracle.password = if (exists("oracle.personal.password")) oracle.personal.password else
    NULL,
  oracle.dbname = if (exists("oracle.personal.server")) oracle.personal.server else NULL
)
}
\description{
allows individual or batch uploading of recapture data
}
