% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.uploading.R
\name{upload_releases}
\alias{upload_releases}
\title{upload_releases}
\usage{
upload_releases(
  username = oracle.personal.user,
  password = oracle.personal.password,
  dbname = oracle.personal.server
)
}
\description{
batch uploads tag release data
}
