% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recapture.uploading.R
\name{batch_upload_recaptures}
\alias{batch_upload_recaptures}
\title{batch_upload_recaptures}
\usage{
batch_upload_recaptures(
  username = oracle.personal.user,
  password = oracle.personal.password,
  dbname = oracle.personal.server
)
}
\description{
batch uploads tag recaptures data
}
