% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recapture.uploading.R
\name{batch_upload_recaptures}
\alias{batch_upload_recaptures}
\title{batch_upload_recaptures}
\usage{
batch_upload_recaptures(
  db = "local",
  oracle.user = oracle.personal.user,
  oracle.password = oracle.personal.password,
  oracle.dbname = oracle.personal.server,
  backups = T
)
}
\description{
batch uploads tag recaptures data
}
