% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recapture.uploading.R
\name{batch_upload_recaptures}
\alias{batch_upload_recaptures}
\title{batch_upload_recaptures}
\usage{
batch_upload_recaptures(
  db = NULL,
  backups = T,
  oracle.user = if (exists("oracle.personal.user", inherits = T)) oracle.personal.user
    else NULL,
  oracle.password = if (exists("oracle.personal.password", inherits = T))
    oracle.personal.password else NULL,
  oracle.dbname = if (exists("oracle.personal.server", inherits = T))
    oracle.personal.server else NULL
)
}
\description{
batch uploads tag recaptures data
}
