% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_deletion.R
\name{delete_recaptures}
\alias{delete_recaptures}
\title{delete_recaptures}
\usage{
delete_recaptures(
  username = oracle.personal.user,
  password = oracle.personal.password,
  dbname = oracle.personal.server
)
}
\description{
allows user to delete chosen tag recaptures and associated paths
}
