% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_deletion.R
\name{delete_recaptures}
\alias{delete_recaptures}
\title{delete_recaptures}
\usage{
delete_recaptures(
  db = NULL,
  oracle.user = oracle.personal.user,
  oracle.password = oracle.personal.password,
  oracle.dbname = oracle.personal.server
)
}
\description{
allows user to delete chosen tag recaptures and associated paths
}
