% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_deletion.R
\name{delete_recaptures}
\alias{delete_recaptures}
\title{delete_recaptures}
\usage{
delete_recaptures(
  db = NULL,
  oracle.user = if (exists("oracle.lobtag.user")) oracle.lobtag.user else NULL,
  oracle.password = if (exists("oracle.lobtag.password")) oracle.lobtag.password else
    NULL,
  oracle.dbname = if (exists("oracle.lobtag.server")) oracle.lobtag.server else NULL
)
}
\description{
allows user to delete chosen tag recaptures and associated paths
}
