% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.generation.R
\name{generate_maps}
\alias{generate_maps}
\title{generate_maps}
\usage{
generate_maps(
  db = NULL,
  people = NULL,
  all.people = FALSE,
  tags = NULL,
  all.tags = FALSE,
  only.unrewarded = FALSE,
  map.token = mapbox.token,
  output.location = NULL,
  max.pixels = 8e+05,
  map.res = 0.9,
  inset.option = T,
  oracle.user = if (exists("oracle.personal.user")) oracle.personal.user else NULL,
  oracle.password = if (exists("oracle.personal.password")) oracle.personal.password else
    NULL,
  oracle.dbname = if (exists("oracle.personal.server")) oracle.personal.server else NULL
)
}
\description{
creates maps of tag movement for participants
}
