% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.generation.R
\name{generate_maps}
\alias{generate_maps}
\title{generate_maps}
\usage{
generate_maps(
  map.token = mapbox.token,
  output.location = output.dir,
  people = NULL,
  all.people = FALSE,
  db = "local",
  oracle.user = oracle.personal.user,
  oracle.password = oracle.personal.password,
  oracle.dbname = oracle.personal.server
)
}
\description{
creates maps of tag movement for participants
}
