% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.generation.R
\name{generate_paths}
\alias{generate_paths}
\title{generate_paths}
\usage{
generate_paths()
}
\description{
Uses releases and recapture data with spatial/depth information to draw plausible paths of animal movement
}
