% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.generation.R
\name{generate_paths}
\alias{generate_paths}
\title{generate_paths}
\usage{
generate_paths(
  db = NULL,
  oracle.user = oracle.personal.user,
  oracle.password = oracle.personal.password,
  oracle.dbname = oracle.personal.server,
  tags = "all",
  depth.raster.path = system.file("data", "gebco_2024.tif", package = "LobTag2"),
  neighborhood = 8,
  type = "least.cost",
  regen.paths = FALSE
)
}
\description{
Uses releases and recapture data with spatial/depth information to draw plausible paths of animal movement
}
