% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.generation.R
\name{generate_paths}
\alias{generate_paths}
\title{generate_paths}
\usage{
generate_paths(
  depth.raster.path = "C:/bio.data/bio.lobster/data/tagging/depthraster2.tif",
  neighborhood = 16,
  type = "least.cost"
)
}
\description{
Uses releases and recapture data with spatial/depth information to draw plausible paths of animal movement
}
