% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.generation.R
\name{generate_paths}
\alias{generate_paths}
\title{generate_paths}
\usage{
generate_paths(
  db = NULL,
  tags = "all",
  depth.raster.path = system.file("data", "gebco_2024.tif", package = "LobTag2"),
  neighborhood = 8,
  type = "least.cost",
  regen.paths = FALSE,
  save.table = FALSE,
  add.land = T,
  check.land = F,
  oracle.user = if (exists("oracle.lobtag.user")) oracle.lobtag.user else NULL,
  oracle.password = if (exists("oracle.lobtag.password")) oracle.lobtag.password else
    NULL,
  oracle.dbname = if (exists("oracle.lobtag.server")) oracle.lobtag.server else NULL
)
}
\description{
Uses releases and recapture data with spatial/depth information to draw plausible paths of animal movement
}
