% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobster.letters.R
\name{lobster.letters}
\alias{lobster.letters}
\title{lobster.letters}
\usage{
lobster.letters(
  people = NULL,
  db = "Oracle",
  only.unrewarded = T,
  output.location = NULL,
  max.pixels = 4e+05,
  map.res = 0.9,
  oracle.user = if (exists("oracle.lobtag.user")) oracle.lobtag.user else NULL,
  oracle.password = if (exists("oracle.lobtag.password")) oracle.lobtag.password else
    NULL,
  oracle.dbname = if (exists("oracle.lobtag.server")) oracle.lobtag.server else NULL
)
}
\description{
create reward letters for fishers who report tags, letters contain map and text.

Letter text for rewards markdown
}
