% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobster.letters.R
\name{lobster.letters}
\alias{lobster.letters}
\title{lobster.letters}
\usage{
lobster.letters(region = "ScotianShelf", db = "Oracle", output.location = NULL)
}
\description{
create reward letters for fishers who report tags, letters contain map and text.

Letter text for rewards markdown
}
