% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.generation.R
\name{map_by_factor}
\alias{map_by_factor}
\title{map_by_factor}
\usage{
map_by_factor(
  db = NULL,
  filter.from = NULL,
  filter.by = NULL,
  filter.for = NULL,
  map.by = NULL,
  factor.by = NULL,
  all.releases = F,
  show.releases = T,
  show.recaptures = T,
  tag.prefix = NULL,
  add.paths = F,
  map.token = mapbox.token,
  max.pixels = 8e+05,
  map.res = 0.9,
  inset.map = F,
  inset.option = T,
  zoom.out = 1,
  point.size = 1.5,
  file.type = "pdf",
  output.location = NULL,
  dpi = 900,
  oracle.user = if (exists("oracle.lobtag.user")) oracle.lobtag.user else NULL,
  oracle.password = if (exists("oracle.lobtag.password")) oracle.lobtag.password else
    NULL,
  oracle.dbname = if (exists("oracle.lobtag.server")) oracle.lobtag.server else NULL
)
}
\description{
general mapping function for mapping releases and returns by custom factor
}
