% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.generation.R
\name{map_by_factor}
\alias{map_by_factor}
\title{map_by_factor}
\usage{
map_by_factor(
  db = NULL,
  factor.from = NULL,
  map.by = NULL,
  filter.maps.for = NULL,
  group.by = NULL,
  all.releases = F,
  show.releases = T,
  show.recaptures = T,
  tag.prefix = NULL,
  add.paths = F,
  map.token = mapbox.token,
  output.location = NULL,
  set.output = T,
  set.inset = T,
  max.pixels = 8e+05,
  map.res = 0.9,
  inset.map = T,
  point.size = 1.5,
  file.type = "pdf",
  margin = 1,
  oracle.user = if (exists("oracle.personal.user")) oracle.personal.user else NULL,
  oracle.password = if (exists("oracle.personal.password")) oracle.personal.password else
    NULL,
  oracle.dbname = if (exists("oracle.personal.server")) oracle.personal.server else NULL
)
}
\description{
specific mapping function for mapping releases and returns by custom factor
}
