% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobster.letters.R
\name{send.lobster.letters}
\alias{send.lobster.letters}
\title{send.lobster.letters}
\usage{
send.lobster.letters(
  db = "Oracle",
  auto.authenticate = T,
  oracle.user = oracle.lobtag.user,
  oracle.password = oracle.lobtag.password,
  oracle.dbname = oracle.lobtag.server
)
}
\description{
Review and send reward letters to fishers emails.
}
