% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recapture.uploading.R
\name{upload_recaptures}
\alias{upload_recaptures}
\title{upload_recaptures}
\usage{
upload_recaptures(
  db = "local",
  oracle.user = oracle.personal.user,
  oracle.password = oracle.personal.password,
  oracle.dbname = oracle.personal.server,
  backups = T
)
}
\description{
allows individual or batch uploading of recapture data
}
