% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recapture.uploading.R
\name{upload_recaptures}
\alias{upload_recaptures}
\title{upload_recaptures}
\usage{
upload_recaptures(
  username = oracle.personal.user,
  password = oracle.personal.password,
  dbname = oracle.personal.server
)
}
\description{
allows individual or batch uploading of recapture data
}
