% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.uploading.R
\name{upload_releases}
\alias{upload_releases}
\title{upload_releases}
\usage{
upload_releases(
  db = NULL,
  oracle.user = if (exists("oracle.personal.user", inherits = T)) oracle.personal.user
    else NULL,
  oracle.password = if (exists("oracle.personal.password", inherits = T))
    oracle.personal.password else NULL,
  oracle.dbname = if (exists("oracle.personal.server", inherits = T))
    oracle.personal.server else NULL
)
}
\description{
batch uploads tag release data
}
