% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.uploading.R
\name{upload_releases}
\alias{upload_releases}
\title{upload_releases}
\usage{
upload_releases(
  db = "local",
  oracle.user = oracle.personal.user,
  oracle.password = oracle.personal.password,
  oracle.dbname = oracle.personal.server
)
}
\description{
batch uploads tag release data
}
