% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.uploading.R
\name{upload_releases}
\alias{upload_releases}
\title{upload_releases}
\usage{
upload_releases(
  db = NULL,
  overwrite.tags = F,
  oracle.user = if (exists("oracle.lobtag.user", inherits = T)) oracle.lobtag.user else
    NULL,
  oracle.password = if (exists("oracle.lobtag.password", inherits = T))
    oracle.lobtag.password else NULL,
  oracle.dbname = if (exists("oracle.lobtag.server", inherits = T)) oracle.lobtag.server
    else NULL
)
}
\description{
batch uploads tag release data
}
